backend/requirements.txt
# ============================================================
# ERP System - Backend Dependencies
# Python 3.x | Django 6.x
# ============================================================

# --- Core Framework ---
Django==6.0.3
djangorestframework==3.17.0
asgiref==3.11.1

# --- Authentication ---
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1

# --- CORS ---
django-cors-headers==4.9.0

# --- Database ---
psycopg2-binary==2.9.11
sqlparse==0.5.5

# --- Filtering ---
django-filter==25.2

# --- Async / Task Queue ---
celery==5.6.2
redis==7.3.0
kombu==5.6.2
billiard==4.2.4
amqp==5.3.1
vine==5.1.0

# --- Celery Beat (scheduled tasks) ---
django-celery-beat==2.9.0
django_celery_results==2.6.0
django-timezone-field==7.2.1
python-crontab==3.3.0
cron-descriptor==1.4.5

# --- WebSocket / Realtime ---
channels==4.3.2
channels_redis==4.3.0
daphne==4.2.1

# --- ASGI / WebSocket protocol ---
autobahn==25.12.2
Twisted==25.5.0
txaio==25.12.2

# --- File Handling ---
Pillow==12.1.1
openpyxl==3.1.5

# --- Environment Config ---
python-decouple==3.8
python-dotenv==1.0.0

# --- ZKTeco Device ---
pyzk==0.9
pyserial==3.5

# --- Dev Utilities ---
django-extensions==4.1

# --- Misc Dependencies ---
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
python-dateutil==2.9.0.post0
six==1.17.0
tzdata==2025.3
tzlocal==5.3.1
packaging==26.0
attrs==26.1.0
cffi==2.0.0
cryptography==46.0.5
pyOpenSSL==26.0.0
service-identity==24.2.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
Incremental==24.11.0
constantly==23.10.4
hyperlink==21.0.0
Automat==25.4.16
zope.interface==8.2
idna==3.11
msgpack==1.1.2
cbor2==5.8.0
py-ubjson==0.16.1
ujson==5.12.0
pycparser==3.0
typing_extensions==4.15.0
prompt_toolkit==3.0.52
wcwidth==0.6.0
